requirements.txt
pytube==12.1.0
python-dotenv==0.20.0